+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Simplify exception hierachy - just one type `MemcacheError` now for
 * Remove `defaultOptions` and `defaultServerSpec`, will revist usefulness.
   exceptions.
+* Remove many `Typeable` instances.
+* Support better testing with a mock Memcached server.
 * Greatly improve documentation.
 
 # 0.1.0.1 - February 26th, 2016